requirements.txt
asgiref==3.5.0
astroid==2.11.2
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.4
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.5
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.0
packaging==23.0
pep8==1.7.1
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.4
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.0
